+++ vars/plWindows.groovy
@@ -7,8 +7,9 @@ def call(duration = 1, unit = "HOURS", Closure body) {
   node('windows') {
     duration = duration ?: 1
     unit = unit ?: 'HOURS'
-    assert (duration instanceof Integer) : "duration must be a Integer"
-    assert (unit instanceof String) : "unit must be a String"
+    assert (duration instanceof Integer)  : "WorkflowLibs :: plWindows :: duration must be a Integer"
+    assert (unit instanceof String)       : "WorkflowLibs :: plWindows :: unit must be a String"
+
     timeout(time: duration, unit: unit.toUpperCase()) {
       ansiColor('xterm') {
         concurPipeline.debugPrint(['duration' : duration, 'unit' : unit])